+++ src/helpers/scripts.js
@@ -1,4 +1,4 @@
-import { getFirestoreCollection, setFirestoreDoc, updateFirestoreDoc } from '/src/crud.js'
+import { getFirestoreCollection, setFirestoreDoc, updateFirestoreDoc } from '/src/helpers/crud.js'
 import { 
   createIndividualFirestoreDocForEachTaskInAllTasks, 
   applyFuncToEveryTreeNode,
