+++ src/lib/ExperimentalCanvas.svelte
@@ -9,7 +9,7 @@
 <script>
   import { onMount } from 'svelte'
   import ExperimentalCanvasWhiteboard from '$lib/ExperimentalCanvasWhiteboard.svelte'
-  import { setFirestoreDoc } from '/src/crud.js'
+  import { setFirestoreDoc } from '/src/helpers/crud.js'
   import { getRandomID } from '/src/helpers/everythingElse.js'
   import { user } from '/src/store.js'
