+++ src/lib/ManageReusableTasksWeeklyPopup.svelte
@@ -97,7 +97,7 @@
     createFirestoreQuery, 
     getFirestoreQuery, 
     updateFirestoreDoc 
-  } from '/src/crud.js'
+  } from '/src/helpers/crud.js'
   import { user } from '/src/store.js'
   import { onMount, onDestroy } from 'svelte'
   import ExperimentalCanvas from '$lib/ExperimentalCanvas.svelte'
